requirements.api.txt
fastapi==0.117.1
uvicorn==0.37.0
python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
pandas==2.3.2
streamlit==1.39.0
plotly==5.24.0
requests==2.32.5
beautifulsoup4==4.13.5
lxml==6.0.2
python-multipart==0.0.20
# watchdog==6.0.0